misc: move coercion nad date check to appropriate modules
